Added checkout function
