--- lib/components/Unusable/drawer.dart
@@ -1,83 +1,0 @@
-// import 'package:project2/pages/settings.dart';
-// import 'package:project2/services/auth/authService.dart';
-// import 'package:flutter/material.dart';
-
-// class MyDrawer extends StatelessWidget {
-//   const MyDrawer({super.key});
-
-//   void logout() {
-//     final auth = AuthService();
-//     auth.signOut();
-//   }
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Drawer(
-//       backgroundColor: Theme.of(context).colorScheme.background,
-//       child: Column(
-//         mainAxisAlignment: MainAxisAlignment.spaceBetween,
-//         children: [
-//           //logo
-//           Column(
-//             children: [
-//               DrawerHeader(
-//                   decoration: BoxDecoration(
-//                       color: Theme.of(context).colorScheme.secondary),
-//                   child: Image.asset('lib/images/LogoColored.png',
-//                       width: 320, fit: BoxFit.cover)),
-//               Padding(
-//                 padding: const EdgeInsets.only(left: 25.0),
-//                 child: ListTile(
-//                   title: Text(
-//                     "Home",
-//                     style:
-//                     TextStyle(color: Theme.of(context).colorScheme.primary),
-//                   ),
-//                   leading: Icon(Icons.home,
-//                       color: Theme.of(context).colorScheme.primary),
-//                   onTap: () {
-//                     //pop drawer
-//                     Navigator.pop(context);
-//                   },
-//                 ),
-//               ),
-//               Padding(
-//                 padding: const EdgeInsets.only(left: 25.0),
-//                 child: ListTile(
-//                   title: Text(
-//                     "Settings",
-//                     style:
-//                     TextStyle(color: Theme.of(context).colorScheme.primary),
-//                   ),
-//                   leading: Icon(Icons.settings,
-//                       color: Theme.of(context).colorScheme.primary),
-//                   onTap: () {
-//                     // pop drawer
-//                     Navigator.pop(context);
-//                     // navigate to settings page
-//                     Navigator.push(
-//                       context,
-//                       MaterialPageRoute(builder: (context) => SettingsPage()),
-//                     );
-//                   },
-//                 ),
-//               ),
-//             ],
-//           ),
-//           Padding(
-//             padding: const EdgeInsets.only(left: 25.0, bottom: 25),
-//             child: ListTile(
-//               title: Text(
-//                 "Logout",
-//                 style: TextStyle(color: Theme.of(context).colorScheme.primary),
-//               ),
-//               leading: Icon(Icons.logout,
-//                   color: Theme.of(context).colorScheme.primary),
-//               onTap: () => logout(),
-//             ),
-//           )
-//         ],
-//       ),
-//     );
-//   }
-// }
