--- lib/components/Unusable/foodtile.dart
@@ -1,84 +1,0 @@
-// import 'package:project2/models/food.dart';
-// import 'package:flutter/material.dart';
-
-// class MyFoodTile extends StatelessWidget {
-//   final Food food;
-//   final void Function()? onTap;
-//   const MyFoodTile({super.key, required this.food, required this.onTap});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Column(
-//       children: [
-//         GestureDetector(
-//           onTap: onTap,
-//           child: Container(
-//             decoration: BoxDecoration(
-//               color: Colors.white,
-//               borderRadius: BorderRadius.circular(20),
-//               boxShadow: [
-//                 BoxShadow(
-//                   color: Colors.grey.withOpacity(0.3),
-//                   spreadRadius: 2,
-//                   blurRadius: 5,
-//                   offset: Offset(0, 3),
-//                 ),
-//               ],
-//             ),
-//             margin: EdgeInsets.symmetric(vertical: 10, horizontal: 15),
-//             padding: EdgeInsets.all(15),
-//             child: Row(
-//               children: [
-//                 Expanded(
-//                   child: Column(
-//                     crossAxisAlignment: CrossAxisAlignment.start,
-//                     children: [
-//                       Text(
-//                         food.name,
-//                         style: TextStyle(
-//                           fontWeight: FontWeight.bold,
-//                           fontSize: 18,
-//                           color: Theme.of(context).colorScheme.primary,
-//                         ),
-//                       ),
-//                       SizedBox(height: 5),
-//                       Text(
-//                         "\$" + food.price.toString(),
-//                         style: TextStyle(
-//                           color: Colors.green,
-//                           fontWeight: FontWeight.bold,
-//                           fontSize: 16,
-//                         ),
-//                       ),
-//                       SizedBox(height: 10),
-//                       Text(
-//                         food.description,
-//                         style: TextStyle(
-//                           color: Colors.grey[600],
-//                         ),
-//                       ),
-//                     ],
-//                   ),
-//                 ),
-//                 SizedBox(width: 10),
-//                 ClipRRect(
-//                   borderRadius: BorderRadius.circular(15),
-//                   child: Image(
-//                     image: AssetImage(food.imagePath),
-//                     height: 120,
-//                   ),
-//                 ),
-//               ],
-//             ),
-//           ),
-//         ),
-//         Divider(
-//           color: Theme.of(context).colorScheme.tertiary,
-//           height: 0,
-//           indent: 15,
-//           endIndent: 15,
-//         ),
-//       ],
-//     );
-//   }
-// }
