+++ lib/pages/checkout.dart
@@ -4,15 +4,16 @@ import 'package:project2/components/bottomnav.dart';
 import 'package:project2/components/loading.dart';
 import 'package:project2/providers/food.dart';
 import 'package:project2/components/receipt.dart';
+import 'package:project2/components/mycheckout.dart';
 
-class TrackPage extends StatefulWidget {
-  const TrackPage({super.key});
+class CheckoutPage extends StatefulWidget {
+  const CheckoutPage({super.key});
 
   @override
-  State<TrackPage> createState() => _TrackPageState();
+  State<CheckoutPage> createState() => _CheckoutPageState();
 }
 
-class _TrackPageState extends State<TrackPage> {
+class _CheckoutPageState extends State<CheckoutPage> {
   FirestoreService db = FirestoreService();
   bool _isLoading = true;
 
@@ -24,9 +25,7 @@ class _TrackPageState extends State<TrackPage> {
       });
     });
     super.initState();
-
     final foodMethods = FoodMethods();
-
     String receipt = foodMethods.displayCartReceipt();
     db.saveOrderToDatabase(receipt);
   }
@@ -37,17 +36,20 @@ class _TrackPageState extends State<TrackPage> {
       appBar: AppBar(
         backgroundColor: Colors.blue[600],
         foregroundColor: Colors.white,
-        title: Text("Track".toUpperCase()),
+        title: Text("Checkout".toUpperCase()),
         centerTitle: true,
         elevation: 4,
         shadowColor: Colors.grey,
       ),
       body: _isLoading
           ? const LoadingWidget()
-          : Column(
-        children: [
-          MyReceipt(),
-        ],
+          : SingleChildScrollView(
+        child: Column(
+          children: [
+            MyCheckout(),
+            MyReceipt(),
+          ],
+        ),
       ),
       bottomNavigationBar: const MyBottomNavBar(currentIndex: 1),
     );
