+++ lib/services/databaseService.dart
@@ -1,14 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class FirestoreService {
-  // final CollectionReference orders = FirebaseFirestore.instance.collection('orders');
+  final CollectionReference orders = FirebaseFirestore.instance.collection('orders');
 
-  // Future<void> saveOrderToDatabase(String receipt) async {
-  //   await orders.add({
-  //     'data' : DateTime.now(),
-  //     'order': receipt,
-  //   });
-  // }
+  Future<void> saveOrderToDatabase(String receipt) async {
+    await orders.add({
+      'data' : DateTime.now(),
+      'order': receipt,
+    });
+  }
 
   final FirebaseFirestore _db = FirebaseFirestore.instance;
