--- lib/services/orderService.dart
@@ -1,12 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-
-class FirestoreService {
-  final CollectionReference orders = FirebaseFirestore.instance.collection('orders');
-
-  Future<void> saveOrderToDatabase(String receipt) async {
-    await orders.add({
-      'data' : DateTime.now(),
-      'order': receipt,
-    });
-  }
-}
